some code quality fixes plus CI
