+++ backend/tests/unit/infrastructure/auth/session/test_api.py
@@ -4,8 +4,17 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from fastapi import Request
 
+from src.infrastructure.auth.session.dependencies import (
+    get_current_session_data,
+    get_current_user,
+    get_session_from_cookie,
+    get_session_manager,
+    verify_csrf_token,
+)
 from src.infrastructure.auth.session.manager import SessionManager
 from src.infrastructure.auth.session.schemas import SessionData
+from src.infrastructure.database.session import async_session
+from src.interfaces.main import app
 
 
 @pytest.mark.asyncio
@@ -15,9 +24,6 @@ async def test_login_endpoint_success(client, test_user, db_session, monkeypatch
     mock_manager.create_session = AsyncMock(return_value=("test-session-id", "test-csrf-token"))
     mock_manager.set_session_cookies = MagicMock()
     mock_manager.track_login_attempt = AsyncMock(return_value=(True, 5))
-
-    from src.infrastructure.auth.session.dependencies import get_session_manager
-    from src.interfaces.main import app
 
     app.dependency_overrides[get_session_manager] = lambda: mock_manager
 
@@ -81,9 +87,6 @@ async def test_logout_endpoint(client, test_user, db_session, monkeypatch):
     mock_manager.terminate_session = AsyncMock(return_value=True)
     mock_manager.clear_session_cookies = MagicMock()
 
-    from src.infrastructure.auth.session.dependencies import get_current_session_data, get_session_manager
-    from src.interfaces.main import app
-
     async def mock_get_current_session_data(request: Request, session_id=None, session_manager=None):
         return session_data
 
@@ -129,9 +132,6 @@ async def test_refresh_csrf_token_endpoint(client, test_user, db_session, monkey
     mock_manager.regenerate_csrf_token = AsyncMock(return_value="new-csrf-token")
     mock_manager.session_timeout = timedelta(minutes=30)  # Add session_timeout attribute
 
-    from src.infrastructure.auth.session.dependencies import get_current_session_data, get_session_manager
-    from src.interfaces.main import app
-
     async def mock_get_current_session_data(request: Request, session_id=None, session_manager=None):
         return session_data
 
@@ -174,10 +174,6 @@ async def test_unauthorized_access(client, test_user, db_session):
 @pytest.mark.asyncio
 async def test_protected_endpoint_access(client, test_user, db_session, monkeypatch):
     """Test accessing protected endpoint with valid session."""
-    from src.infrastructure.auth.session.dependencies import get_current_user, get_session_from_cookie, verify_csrf_token
-    from src.infrastructure.database.session import async_session
-    from src.interfaces.main import app
-
     test_user_with_image = test_user.copy()
     test_user_with_image["profile_image_url"] = "https://example.com/default-avatar.png"
 
@@ -244,9 +240,6 @@ async def test_login_endpoint_with_rate_limiting(client, test_user, db_session, 
     mock_manager.track_login_attempt = AsyncMock(return_value=(True, 4))
     mock_manager.create_session = AsyncMock(return_value=("test-session-id", "test-csrf-token"))
     mock_manager.set_session_cookies = MagicMock()
-
-    from src.infrastructure.auth.session.dependencies import get_session_manager
-    from src.interfaces.main import app
 
     app.dependency_overrides[get_session_manager] = lambda: mock_manager
